Amends to navigation
