+++ src/components/ShipManagement.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { useGame } from '../context/GameContext';
-import { Ship, Equipment, ShipClass, ShipSpeed, ShipRange, EquipmentSlot } from '../types/game';
-import { shipTemplates } from '../data/ships';
+import { Ship, Equipment, EquipmentSlot } from '../types/game';
 import { equipmentList } from '../data/equipment';
 import { Shield, Sword, Pickaxe } from 'lucide-react';
 
@@ -14,34 +13,6 @@ const slotIcons: Record<EquipmentSlot, React.ReactNode> = {
 export const ShipManagement: React.FC = () => {
   const { state, dispatch } = useGame();
   const [selectedShip, setSelectedShip] = useState<Ship | null>(null);
-  const [selectedEquipment, setSelectedEquipment] = useState<Equipment | null>(null);
-
-  const handleUpgradeSpaceDock = () => {
-    dispatch({ type: 'UPGRADE_SPACE_DOCK' });
-  };
-
-  const getShipSpeedColor = (speed: ShipSpeed) => {
-    switch (speed) {
-      case 'fast': return 'text-green-400';
-      case 'medium': return 'text-yellow-400';
-      case 'slow': return 'text-red-400';
-    }
-  };
-
-  const getShipRangeColor = (range: ShipRange) => {
-    switch (range) {
-      case 'long': return 'text-green-400';
-      case 'medium': return 'text-yellow-400';
-      case 'short': return 'text-red-400';
-    }
-  };
-
-  const handleBuyShip = (shipClass: ShipClass) => {
-    const template = shipTemplates[shipClass];
-    if (state.player.credits >= template.cost) {
-      dispatch({ type: 'BUY_SHIP', shipClass });
-    }
-  };
 
   const handleSellShip = (ship: Ship) => {
     if (state.player.ships.length > 1) {
@@ -82,22 +53,6 @@ export const ShipManagement: React.FC = () => {
       <div className="flex justify-between items-center mb-6">
         <h2 className="text-2xl font-bold">Ship Management</h2>
         <div className="text-xl">Credits: {state.player.credits}</div>
-      </div>
-
-      <div className="flex justify-between items-center mb-6">
-        <h2 className="text-2xl font-bold">Ship Management</h2>
-        <div className="flex items-center gap-4">
-          <div className="text-xl">Credits: {state.player.credits}</div>
-          {!state.player.hasAdvancedSpaceDock && (
-            <button
-              onClick={handleUpgradeSpaceDock}
-              disabled={state.player.credits < 300}
-              className="px-4 py-2 bg-purple-600 hover:bg-purple-700 disabled:bg-gray-600 rounded"
-            >
-              Upgrade Space Dock (300 CR)
-            </button>
-          )}
-        </div>
       </div>
 
       {/* Ship List */}
@@ -142,59 +97,6 @@ export const ShipManagement: React.FC = () => {
             )}
           </div>
         ))}
-      </div>
-
-      {/* Ship Purchase */}
-      <div className="mb-6">
-        <h3 className="text-xl font-bold mb-4">Available Ships</h3>
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-          {(Object.entries(shipTemplates) as [ShipClass, Ship][])
-            .filter(([className]) => 
-              state.player.hasAdvancedSpaceDock || 
-              ['prospector', 'assault-fighter', 'scoutship'].includes(className)
-            )
-            .map(([className, template]) => (
-              <div key={className} className="bg-gray-700 p-4 rounded-lg">
-                <div className="flex justify-between items-center mb-2">
-                  <span className="capitalize font-semibold">
-                    {className.replace('-', ' ')}
-                  </span>
-                  <span>{template.cost} credits</span>
-                </div>
-                <div className="space-y-2 mb-4">
-                  <div className="flex gap-4">
-                    {(Object.entries(template.maxEquipmentSlots) as [EquipmentSlot, number][])
-                      .map(([slot, max]) => (
-                        <div key={slot} className="flex items-center gap-1">
-                          {slotIcons[slot]}
-                          <span>{max}</span>
-                        </div>
-                      ))}
-                  </div>
-                  <div className="flex gap-4 text-sm">
-                    <span className={getShipSpeedColor(template.speed)}>
-                      {template.speed} speed
-                    </span>
-                    <span className={getShipRangeColor(template.range)}>
-                      {template.range} range
-                    </span>
-                    {template.carrierCapacity && (
-                      <span className="text-blue-400">
-                        Carries {template.carrierCapacity} ships
-                      </span>
-                    )}
-                  </div>
-                </div>
-                <button
-                  onClick={() => handleBuyShip(className)}
-                  disabled={state.player.credits < template.cost}
-                  className="w-full px-4 py-2 bg-blue-600 hover:bg-blue-700 disabled:bg-gray-600 rounded"
-                >
-                  Buy Ship
-                </button>
-              </div>
-            ))}
-        </div>
       </div>
 
       {/* Equipment Management */}
